IN-200: make post open on only profile page
