+++ src/common/paths.ts
@@ -4,7 +4,6 @@ export const Paths = {
   forgotPassword: '/auth/forgot-password',
   generalInfo: '/settings',
   home: '/home',
-  homePost: '/home/:postId',
   messenger: '/messenger',
   privacyPolicy: '/auth/privacy-policy',
   profile: '/profile',
@@ -13,7 +12,6 @@ export const Paths = {
   },
   profileSettings: '/profile/settings',
   publicMainPage: '/',
-  publicMainPagePost: '/:postId',
   recoveryPassword: '/auth/recovery',
   registrationEmailResending: '/auth/registration-email-resending',
   search: '/search',
@@ -21,5 +19,6 @@ export const Paths = {
   signUp: '/auth/sign-up',
   statistics: '/statistics',
   termsOfService: '/auth/terms-of-service',
-  userProfile: '/profile/:userId/:postId',
+  userPost: '/profile/:userId/:postId',
+  userProfile: '/profile/:userId',
 }
