+++ src/features/newsFeed/ui/PostItem.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
 import {
-  PathService,
   BasicPost,
+  PathService,
   Paths,
   TextUnfolding,
   Typography,
@@ -64,7 +64,10 @@ const PostItem = ({ item, router }: Props) => {
         <CustomNewsFeedPopover />
       </div>
       <Link
-        href={PathService.generatePath(Paths.homePost, { postId: item.id })}
+        href={PathService.generatePath(Paths.userPost, {
+          postId: item.id,
+          userId: item.ownerId,
+        })}
       >
       <section className={'relative w-full aspect-square overflow-hidden mb-3'}>
         {item.images.length > 0 ? (
