+++ src/features/posts/ui/PostModal/PostModal.tsx
@@ -1,17 +1,17 @@
 import React from 'react'
 
 import { Close } from '@/assets/icons'
-import { Nullable, Overlay } from '@/common'
+import { Nullable, Overlay, PathService, Paths } from '@/common'
 import { Post, PostComments } from '@/features/posts'
 import { PostContent } from '@/features/posts/ui/PostModal/Content/PostContent'
+import Link from 'next/link'
 
 type Props = {
   comments: Nullable<PostComments>
-  onClose: () => void
   post: Nullable<Post>
 }
 
-export const PostModal = ({ comments, onClose, post }: Props) => {
+export const PostModal = ({ comments, post }: Props) => {
   return (
     <Overlay
       className={'flex justify-center items-center'}
@@ -22,10 +22,15 @@ export const PostModal = ({ comments, onClose, post }: Props) => {
           comments={comments}
           post={post}
         />
-        <Close
-          className={'absolute w-6 h-6 -top-8 -right-10 cursor-pointer'}
-          onClick={onClose}
-        />
+        <Link
+          href={PathService.generatePath(Paths.userProfile, {
+            userId: post?.ownerId,
+          })}
+        >
+          <Close
+            className={'absolute w-6 h-6 -top-8 -right-10 cursor-pointer'}
+          />
+        </Link>
       </div>
     </Overlay>
   )
