+++ src/features/publicPosts/ui/PublicPosts.tsx
@@ -44,6 +44,7 @@ export const PublicPosts = ({ posts }: Props) => {
                     <Link
                       href={PathService.generatePath(Paths.publicMainPagePost, {
                         postId: post.id,
+                        userId: post.ownerId,
                       })}
                     >
                   <CustomSwiper
