--- src/pages/home/layout.tsx
@@ -1,14 +1,0 @@
-import React, { PropsWithChildren } from 'react'
-
-import { NewsFeed } from '@/features/newsFeed'
-
-const Layout = ({ children }: PropsWithChildren) => {
-  return (
-    <>
-      <NewsFeed />
-      {children}
-    </>
-  )
-}
-
-export default Layout
